Notes on cefpython3
I won't claim a complete of even significant understanding of cefpython3 and its capabilities. What I know now is (hopefully) all the basics/practicalities I'll ever need to know to get cefpython3 to work as a UI renderer.

I have put up a few notes on tutorial.py (a cefpython3 example) that pretty much covers everything you'd need to know so that it's easier to understand when you go through it. Along with the most important problem to be tackled to get Proton up and running.